requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.11.29
chardet==3.0.4
cheroot==6.5.4
CherryPy==18.1.0
cycler==0.10.0
decorator==4.3.2
distlib==0.2.8
Django==2.1.7
dlib==19.16.0
idna==2.8
image==1.5.27
imutils==0.5.2
jaraco.functools==2.0
jsonschema==2.6.0
kiwisolver==1.0.1
matplotlib==3.0.2
more-itertools==6.0.0
numpy==1.16.1
of==1.0.1
opencv-contrib-python==4.0.0.21
opencv-python==4.0.0.21
pandas==0.24.1
Pillow==5.4.1
Pmw==2.0.1
portend==2.3
pymongo==3.7.2
pyparsing==2.3.1
pystray==0.14.4
python-dateutil==2.8.0
python-xlib==0.25
pytz==2018.9
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
tempora==1.14
urllib3==1.24.1
ws4py==0.5.1
zc.lockfile==1.4
